Reestruturação da arvore de pastas do projeto
