+++ src/components/Forms/formSignup/Form.jsx
@@ -1,9 +1,9 @@
 import styles from './Form.module.css'
-import Input from '../layout/input/Input'
-import SubmitButton from '../layout/submitbuttun/SubmitButton'
-import LinkButton from '../layout/linkbutton/LinkButton'
+import Input from '../../layout/input/Input'
+import SubmitButton from '../../layout/submitbuttun/SubmitButton'
+import LinkButton from '../../layout/linkbutton/LinkButton'
 import { useContext, useState } from 'react'
-import { AuthContext } from '../contexts/Auth'
+import { AuthContext } from '../../contexts/Auth'
 
 const Form = () => {
     const { signup } = useContext(AuthContext)
