+++ src/components/Modals/modal_filter_representative/Modal.jsx
@@ -2,7 +2,7 @@ import React, { Fragment } from 'react'
 import { DatePicker, Input } from 'antd';
 
 import style from './Modal.module.css'
-import SubmitButton from '../layout/submitbuttun/SubmitButton';
+import SubmitButton from '../../layout/submitbuttun/SubmitButton';
 
 import { HiXMark } from "react-icons/hi2";
 import { IconContext } from 'react-icons';
